TWINFACES-128 Updates entity icons (#224)
[TWINFACES-289] feat: add context and refactor: some hooks

[TWINFACES-289] feat: add layout and basic tab general

[TWINFACES-289] feat: add factory field in tab general

[TWINFACES-289] feat: add some field in tab general

[TWINFACES-289] feat: implement functional to edit fields

[TWINFACES-328] feat: add steps tab for pipelines

[TWINFACES-289] fix: twin class filtering by head and extends after changing scheme

[TWINFACES-328] feat: resolve PR comments
