+++ src/entities/factory-pipeline-step/libs/hooks/use-filters.ts
@@ -5,6 +5,7 @@ import { useFactoryPipelineSelectAdapter } from "@/entities/factory-pipeline";
 import { useFeaturerSelectAdapter } from "@/entities/featurer";
 import {
   FilterFeature,
+  isPopulatedArray,
   mapToChoice,
   toArray,
   toArrayOfString,
@@ -13,74 +14,86 @@ import {
 import { z } from "zod";
 import { PipelineStepFilterKeys, PipelineStepFilters } from "../../api";
 
-export function usePipelineStepFilters(): FilterFeature<
-  PipelineStepFilterKeys,
-  PipelineStepFilters
-> {
-  const factorySelectAdapter = useFactorySelectAdapter();
-  const factoryPipelineSelectAdapter = useFactoryPipelineSelectAdapter();
-  const featurerSelectAdapter = useFeaturerSelectAdapter(23);
-  const factoryConditionSetSelectAdapter =
-    useFactoryConditionSetSelectAdapter();
+export function usePipelineStepFilters({
+  enabledFilters,
+}: {
+  enabledFilters?: PipelineStepFilterKeys[];
+}): FilterFeature<PipelineStepFilterKeys, PipelineStepFilters> {
+  const fAdapter = useFactorySelectAdapter();
+  const fpAdapter = useFactoryPipelineSelectAdapter();
+  const featurerAdapter = useFeaturerSelectAdapter(23);
+  const fcsAdapter = useFactoryConditionSetSelectAdapter();
+
+  const allFilters: Record<PipelineStepFilterKeys, AutoFormValueInfo> = {
+    idList: {
+      type: AutoFormValueType.tag,
+      label: "Id",
+      schema: z.string().uuid("Please enter a valid UUID"),
+      placeholder: "Enter UUID",
+    },
+    factoryIdList: {
+      type: AutoFormValueType.combobox,
+      label: "Factory",
+      multi: true,
+      ...fAdapter,
+    },
+    factoryPipelineIdList: {
+      type: AutoFormValueType.combobox,
+      label: "Pipeline",
+      multi: true,
+      ...fpAdapter,
+    },
+    factoryConditionSetIdList: {
+      type: AutoFormValueType.combobox,
+      label: "Condition set",
+      multi: true,
+      ...fcsAdapter,
+    },
+    conditionInvert: {
+      type: AutoFormValueType.boolean,
+      label: "Condition invert",
+      hasIndeterminate: true,
+      defaultValue: "indeterminate",
+    },
+    active: {
+      type: AutoFormValueType.boolean,
+      label: "Active",
+      hasIndeterminate: true,
+      defaultValue: "indeterminate",
+    },
+    optional: {
+      type: AutoFormValueType.boolean,
+      label: "Optional",
+      hasIndeterminate: true,
+      defaultValue: "indeterminate",
+    },
+    fillerFeaturerIdList: {
+      type: AutoFormValueType.combobox,
+      label: "Filler featurer",
+      multi: true,
+      ...featurerAdapter,
+    },
+    descriptionLikeList: {
+      type: AutoFormValueType.tag,
+      label: "Description",
+    },
+  };
 
   function buildFilterFields(): Record<
     PipelineStepFilterKeys,
     AutoFormValueInfo
   > {
-    return {
-      idList: {
-        type: AutoFormValueType.tag,
-        label: "Id",
-        schema: z.string().uuid("Please enter a valid UUID"),
-        placeholder: "Enter UUID",
-      },
-      factoryIdList: {
-        type: AutoFormValueType.combobox,
-        label: "Factory",
-        multi: true,
-        ...factorySelectAdapter,
-      },
-      factoryPipelineIdList: {
-        type: AutoFormValueType.combobox,
-        label: "Pipeline",
-        multi: true,
-        ...factoryPipelineSelectAdapter,
-      },
-      factoryConditionSetIdList: {
-        type: AutoFormValueType.combobox,
-        label: "Condition set",
-        multi: true,
-        ...factoryConditionSetSelectAdapter,
-      },
-      conditionInvert: {
-        type: AutoFormValueType.boolean,
-        label: "Condition invert",
-        hasIndeterminate: true,
-        defaultValue: "indeterminate",
-      },
-      active: {
-        type: AutoFormValueType.boolean,
-        label: "Active",
-        hasIndeterminate: true,
-        defaultValue: "indeterminate",
-      },
-      optional: {
-        type: AutoFormValueType.boolean,
-        label: "Optional",
-        hasIndeterminate: true,
-        defaultValue: "indeterminate",
-      },
-      fillerFeaturerIdList: {
-        type: AutoFormValueType.combobox,
-        label: "Filler featurer",
-        multi: true,
-        ...featurerSelectAdapter,
-      },
-      descriptionLikeList: {
-        type: AutoFormValueType.tag,
-        label: "Description",
-      },
-    };
+    if (isPopulatedArray(enabledFilters)) {
+      return enabledFilters.reduce(
+        (filters, key) => {
+          filters[key] = allFilters[key];
+          return filters;
+        },
+        {} as Record<PipelineStepFilterKeys, AutoFormValueInfo>
+      );
+    }
+
+    return allFilters;
   }
 
   function mapFiltersToPayload(
