+++ src/entities/permission-schema/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./resource-link";
+export * from "./permission-schema-icon";
