+++ src/widgets/tables/index.ts
@@ -1,3 +1,4 @@
 export * from "./twin-class-statuses";
 export * from "./twin-class-fields";
 export * from "./datalist-options";
+export * from "./pipeline-steps";
